requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-multipart==0.0.6
jinja2==3.1.3
itsdangerous==2.1.2
bcrypt==4.1.2
python-dotenv==1.0.0
starlette==0.35.1